publishing path image to ROS
